reorganize scripts folder
