+++ hardhat.config.ts
@@ -7,7 +7,7 @@ import 'hardhat-contract-sizer'
 import './scripts/deploy'
 import './scripts/wallet'
 import './scripts/tasks/protocol/mintPlatformIdForAddress'
-import './scripts/mintTalentLayerId'
+import './scripts/tasks/protocol/mintTalentLayerId'
 import './scripts/tasks/protocol/addArbitrator'
 import './scripts/tasks/protocol/removeArbitrator'
 import { Network } from './scripts/config'
